+++ R/edit.R
@@ -95,12 +95,22 @@ edit_r_makevars <- function(scope = c("user", "project")) {
 
 #' @export
 #' @rdname edit
-#' @param type Snippet type. One of: "R", "markdown", "C_Cpp", "Tex",
-#'   "Javascript", "HTML", "SQL"
-edit_rstudio_snippets <- function(type = "R") {
+#' @param type Snippet type (case insensitive text).
+edit_rstudio_snippets <- function(type = c("r", "markdown", "c_cpp", "css",
+  "html", "java", "javascript", "python", "sql", "stan", "tex")) {
   # RStudio snippets stored using R's definition of ~
   # https://github.com/rstudio/rstudio/blob/4febd2feba912b2a9f8e77e3454a95c23a09d0a2/src/cpp/core/system/Win32System.cpp#L411-L458
-  path <- path_home_r(".R", "snippets", path_ext_set(tolower(type), "snippets"))
+  type <- tolower(type)
+  type <- match.arg(type)
+
+  path <- path_home_r(".R", "snippets", path_ext_set(type, "snippets"))
+  if (!file_exists(path)) {
+    ui_done("New snippet file at {ui_path(path)}")
+    ui_info(c(
+      "This masks the default snippets for {ui_field(type)}.",
+      "Delete this file and restart RStudio to restore the default snippets."
+    ))
+  }
   edit_file(path)
 }