+++ tests/testthat/test-edit.R
@@ -51,11 +51,16 @@ test_that("edit_r_XXX('user') ensures the file exists", {
 
   edit_r_makevars("user")
   expect_r_file(".R", "Makevars")
-
+  
+  
   edit_rstudio_snippets(type = "R")
   expect_r_file(".R", "snippets", "r.snippets")
   edit_rstudio_snippets(type = "HTML")
   expect_r_file(".R", "snippets", "html.snippets")
+  expect_error(
+    edit_rstudio_snippets("not-existing-type"),
+    regexp = "should be one of"
+  )
 })
 
 test_that("edit_r_profile() respects R_PROFILE_USER", {